Define type alias config.PipedGit for configv1.PipedGit
